添加改进余弦相似度

+++ recommend/src/main/scala/com/yankaizhang/movielikes/recommend/offlineRecommend/ImpCosSim.scala
@@ -71,8 +71,8 @@ object ImpCosSim {
         val ratingSumOfSq1 = row.getAs[Double](3)
         val ratingSum2 = row.getAs[Double](4)
         val size = row.getAs[Int](5)
-        val count1 = row.getAs[Int](6)
-        val count2 = row.getAs[Int](7)
+        val count1 = row.getAs[Long](6)
+        val count2 = row.getAs[Long](7)
 
         val impCosSim = SimilarityMeasures.improvedCosineSimilarity(dotProduct, math.sqrt(ratingSumOfSq1), math.sqrt(ratingSum2), size, count1, count2)
